Multiple cleanups and fixes from Ihno Krumreich bugs - fork08 -- the exit value of the programm is not calculated correct. - sysctl03 -- the last parameter of sysctl is a value, not a pointer
64-Bit cleanups
all five patches are fixes for the problem that size(int) != size(void *)
The result of the tests depends on the returnvalue of malloc and signal.
testcases.kernel.mem.mtest07.patch
testcases.kernel.mem.waitpid.patch
testcases.kernel.sched.pthreads.patch
testcases.kernel.syscalls.sigaltstack.patch
testcases.kernel.syscalls.signal.patch

warnings
remove of warnings of implecied declared functions and unused
variables.
Testcases.commands.ade.ldd.patch
testcases.kernel.syscalls.creat.patch
testcases.kernel.syscalls.creat_slink.patch
testcases.kernel.syscalls.exit.patch
testcases.kernel.syscalls.fchdir.patch
testcases.kernel.syscalls.fsync.patch
testcases.kernel.syscalls.getcwd.patch
testcases.kernel.syscalls.getdents.patch
testcases.kernel.syscalls.getgid.patch
testcases.kernel.syscalls.getgroups.patch
testcases.kernel.syscalls.mkdir.patch
testcases.kernel.syscalls.mknod.patch
testcases.kernel.syscalls.mmap.patch
testcases.kernel.syscalls.modify_ldt.patch
testcases.kernel.syscalls.mprotect.patch
testcases.kernel.syscalls.mremap.patch
testcases.kernel.syscalls.open.patch
testcases.kernel.syscalls.pipe.patch
testcases.kernel.syscalls.poll.patch
testcases.kernel.syscalls.pread.patch
testcases.kernel.syscalls.pwrite.patch
testcases.kernel.syscalls.read.patch
testcases.kernel.syscalls.readdir.patch
testcases.kernel.syscalls.readlink.patch
testcases.kernel.syscalls.rename.patch
testcases.kernel.syscalls.rmdir.patch
testcases.kernel.syscalls.sched_setscheduler.patch
testcases.kernel.syscalls.sched_yield.patch
testcases.kernel.syscalls.setitimer.patch
testcases.kernel.syscalls.setpriority.patch
testcases.kernel.syscalls.setrlimit.patch
testcases.kernel.syscalls.settimeofday.patch
testcases.kernel.syscalls.sigaction.patch
testcases.kernel.syscalls.stat.patch
testcases.kernel.syscalls.statfs.patch
testcases.kernel.syscalls.sysinfo.patch
testcases.kernel.syscalls.times.patch
testcases.kernel.syscalls.umask.patch
testcases.kernel.syscalls.uname.patch
testcases.kernel.syscalls.utime.patch
testcases.kernel.syscalls.vfork.patch
testcases.kernel.syscalls.vhangup.patch
testcases.kernel.syscalls.wait4.patch

+++ testcases/kernel/mem/mtest07/mallocstress.c
@@ -255,7 +255,7 @@ void *
 alloc_mem(void * threadnum)
 {
     /* thread N will use growth scheme N mod 4 */
-    int err = allocate_free(num_loop, ((int)threadnum) % 4);
+    int err = allocate_free(num_loop, ((intptr_t)threadnum) % 4);
     fprintf(stdout, 
 	"pid[%d]: allocate_free() returned %d, %s.  Thread exiting.\n",
 	getpid(), err,
@@ -360,7 +360,7 @@ main(int	argc,		/* number of input parameters		      */
         }
         else
         {
-            if ((int)th_status != 0)
+            if ((intptr_t)th_status != 0)
             {
                 fprintf(stderr,
                         "main(): thread [%d] - process exited with errors\n",
